requirements.txt
termcolor==3.1.0
pydantic==2.11.4
google-genai>=1.40.0
playwright==1.52.0
browserbase==1.3.0
rich
pytest
PyYAML>=6.0
